[#46241] Expose `employee_email` and `employee_phone` in `mo_values`
This is a crude first take - we probably need some way to let the Jinja
template specify whether a mobile or landline phone number is desired,
or whether we want the work or private email address. As of now, we
simply use the first phone number or email found in the employee's
addresses.