[#46242] Add optional AD setting `new_ad_user_path`, and use it if set
This allows us to create AD users in a different AD OU than the one
given by the `search_base` AD setting. This is needed to test the setup
in Furesø, where we have write access to only one test OU, but are not
allowed to write to the other AD OUs.

Example configuration:

    ...
    "integrations.ad": [
        {..., "search_base": "dc=furesoe,dc=kommune", ...}, ...
    ],
    ...
    "integrations.ad_writer.new_ad_user_path":
        "ou=os2mo,dc=furesoe,dc=kommune",
    ...

This should mean that calls to "New-ADUser" use a "-Path" argument of
"ou=os2mo,dc=furesoe,dc=kommune", which is the path to the AD OU that we
are allowed to write to in Furesø.