[#47947] Rework `viborg_eksterne` into slightly more testable code; change engagement type lookup; add test
In Viborg, a bug was encountered when looking up an "engagement type"
via `engagement_type = lc.classes[engv["engagement_type"]]["title"]`.

The bugfix for now is to read the engagement type from
`lc_historic.classes` rather than `lc.classes`.

This MR rewrites most of `viborg_eksterne.py` as a class-based tool.
This is done to ease the testing of the code.

This MR adds a regression test for the issue encountered.