[#48568] Fix bug in `AdMoSync._read_all_mo_users`
`AdMoSync._read_all_mo_users` raises an `IndexError` if it encounters
one or more "empty" users in `self.lc.users`.

(An "empty" user is a key/value pair where the key is a valid UUID, but
the value is an empty list.)

This MR adds a bugfix and a regression test.

This MR also moves some mocks from `test_ad_life_cycle` to `mocks`.