[#51414] Ensure `AdMoSync._edit_user_attrs` is idempotent
This is tested only for the codepath where `AdMoSync` uses `LoraCache`.
(The test uses a testable subclass of `AdMoSync` which in turn uses a
mock `LoraCache` to record any mutations made on the MO employee.)

This changes `AdMoSync._update_users` to iterate over (MO employee, AD
user) pairs, rather than (MO user UUID, AD user) pairs. This is done so
that `_edit_user_attrs` can access the current state of the MO employee.