feat: [#54929] Increase exports_os2sync timeout
OS2sync creates the hierarchy lazily after the first request and returns
it eventually when it is ready on later requests. exports_os2sync retry
until it get the requested hierarchy, but fails if it does not get
within 5 minutes.

In the last few days on Viborg, we have seen it takes longer than 5
minutes for os2sync to create the hierarchy and exports_os2sync timesout
and fails. We increase the timeout in exports_os2sync to 10min.