fix: [#55014] Don't backup non-existent rundb
In salt-automation commit 15d72913[1], the
`integrations.SD_Lon.import.run_db` setting was promoted to a global
variable, having previously only been set in environments with the
SD_Lon integration enabled. The job-runner, however, uses the mere
existence of this variable to decide if the sqlite database file in the
path should be backed up. The file only exists if the integration is
run, causing the backup to fail in all other environments.

[1] https://git.magenta.dk/labs/salt-automation/-/commit/15d72913