fix: [#54083] `ADWriter._rename_ad_user`: alter sleep from 10 secs to 1 sec
When running the initial `mo_to_ad_sync` at a specific customer, it was
observed that many AD users needed to be renamed.

Sleeping for 10 seconds for each rename operation meant that
`mo_to_ad_sync` would take needlessly long.

It seems the sleep is wholly redundant, but as a first change, reduce
the sleep to 1 second.