fix: [#54774] removed initial class called 'Tilknytningsrolle' for AAK los import
The class was meant as temp. while waiting for aak to supply proper CSV files, which they do now.
AAK then complained about the class not being removed, which it should, but we have a bigger
issue related to why this is. This is a 'hack' which don't involve a lot of changes