feat: [#55238] AD: make username even if surname has no consonants
Removing the line:
```
assert next_consonant, "next name parts must contain at least one
consonant"
```
means that `UserNameGenPermutation` is able to generate usernames even
for names where the surname part(s) contain *no* consonants.

But removing the assert also means that the code can enter an infinite
loop on very degenerate inputs (such as `["A"]`.) To counter this, an
iteration counter is added, and if the number of max iterations is
reached, a `ValueError` is raised, to prevent the username generator
from hanging in an infinite loop.