fix: [#47213] `MOGraphqlSource` supports "self managers"
In Opus environments, employees who are also managers always have a
primary engagement in the org unit where they are managers.

This means that we would set the corresponding AD user as its own
`Manager`, which is undesirable.

This workaround simply returns None in the cases where the given
employee UUID equals the potential return value of
`MOGraphqlSource.get_manager_uuid`.

fix: [#47213] Support "self managers" in `MOGraphqlSource`

This reimplements most of `MOGraphqlSource` and thus the tests are also
rewritten.

The code now requests all org units, along with their engagements and
managers, and constructs the manager map based on that data. This
assumes that all engagements must point to an org unit.

The code uses v7 of the GraphQL API and takes advantage of pagination.

The code considers the validity of the engagements, and only considers
current or future engagements as valid.

To keep the contract of `get_manager_uuid`, the implementation returns
None in case the employee UUID cannot be looked up in the manager map.