fix: [#57295] chunk sql export sql commits
Originally, the SQL exporter wrote all items in each collection before
committing. This required a lot of memory, so was replaced with
per-object commit in 865bf3a26de1aad448221a46dea2216bec7d9753. This
requires a lot of ressources on the SQL server. As a middle-ground, this
commit introduces chunked committing.