chore(LoraCache): [#62459] don't take graphql session as argument
The caller doesn't know which version of GraphQL we are going to use.
Normally, in other programs, it makes sense to initialise the session in
one place (FastRAMQPI init) and pass it around, but since the LoRaCache
is used as a library, the callers cannot initialise the session for us.